Add a function to read the BMC version.
Removing the SPI padding is quite clunky. As is ensuring you only pass exactly the right amount of data to the decoder. Which means knowing how many bytes you asked for, and whether the result was a success or not.

If it's not a success (i.e. you read a bad register) do you get 2 bytes back, or 2 + N bytes? Probably only 2!