Store vga code and font data in RAM, as flash is too slow
It looks like the pixel generating loop is nearly exhausting the time
budget, and there is no margin left for flash cache misses.

I had to increase the RAM reserved for the BIOS from 16k to 24k.
This might be excessive. Perhaps it is possibe to extract a smaller
selection of really timing critical code and move only that to RAM?