Handle BMC interrupts.
Sets up the MCP23S17 to provide a level low when any IRQ pin is low. Track the pin state in an interrupt handler (which watches both edges), then only poll the BMC for keyboard bytes when the flag is set.

Currently panics on startup because we can't read back out of the MCP23S17. Unsure why. Maybe the BMC is holding CIPO low?