Rework BMC read function.
We no longer store all the padding. Instead we poll the BMC until it gives us a valid header byte. Then we transfer the remainder of the message in one go.

Should avoid the issue of the retry warning on the BMC slowing it down so much that the retries always failed.