Cleaning up render loop.
Trying to remove static mut vars, and keep things self-contained.

Currently it boots in 80x25 but when it switches to 80x30 the rendering gets messed up. Probably need to sync the render loop with the frame loop somehow. Maybe set a flag when V-Blank starts, so we can do some per-frame processing.